Sandbox is STEPIK_MODULE_TYPE now
